ajout de math et modification d.but modification du parseur

+++ HLHML/LanguageElements/Adjectifs/PlusGrand.cs
@@ -19,7 +19,7 @@ namespace HLHML.LanguageElements.Adjectifs
                 throw new InvalidNodeNumberException($"L'adjectif plus 'plus petit que' doit avoir deux noeuds enfants. Celui-ci en à {Childs.Count}.");
             }
 
-            return int.Parse(NodeVisitor.Eval(Childs[0])) > int.Parse(NodeVisitor.Eval(Childs[1]));
+            return double.Parse(NodeVisitor.Eval(Childs[0])) > double.Parse(NodeVisitor.Eval(Childs[1]));
         }
     }
 }
